Refactor: moved controllers to own folder
